pyproject.toml
[project]
name = "nanocopaw"
dynamic = ["version"]
description = "nanoCoPaw is a binary-first, channel-only AI assistant. Configure on launch, talk through DingTalk, Feishu, QQ, Discord, and iMessage."
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "discord-py>=2.3",
    "dingtalk-stream>=0.24.3",
    "uvicorn>=0.40.0",
    "apscheduler>=3.11.2,<4",
    "playwright>=1.49.0",
    "questionary>=2.1.1",
    "python-dotenv>=1.0.0",
    "lark-oapi>=1.5.3",
    "pydantic>=2.0.0",
]

[tool.setuptools.dynamic]
version = {attr = "copaw.__version__.__version__"}

[tool.setuptools]
packages = { find = { where = ["src"] } }
include-package-data = true

[tool.setuptools.package-data]
"copaw" = [
    "skills/**",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
nanocopaw = "copaw.launcher:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "pre-commit>=4.2.0",
    "pytest-cov>=6.2.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
